Restructure support files
