+++ frontend/cypress/integration/actions_spec.ts
@@ -1,12 +1,12 @@
-import { EvaluationSeed } from '../support/evaluation_seed'
+import { EvaluationSeed } from '../support/testsetup/evaluation_seed'
 import { Progression, Priority, Role } from '../../src/api/models'
-import { FUSION_DATE_LOCALE } from '../support/helpers'
+import { FUSION_DATE_LOCALE } from '../support/helpers/helpers'
 import { barrierToString } from '../../src/utils/EnumToString'
-import { Action, Note, Participant } from '../support/mocks'
+import { Action, Note, Participant } from '../support/testsetup/mocks'
 import { ActionsGrid, ActionsTab, CreateActionDialog, EditActionDialog, mapPriority } from '../page_objects/action'
 import { EvaluationPage, QuestionnaireSidePanel } from '../page_objects/evaluation'
 import { ConfirmationDialog, DropdownSelect } from '../page_objects/common'
-import { DELETE_ACTION } from '../support/gql'
+import { DELETE_ACTION } from '../support/testsetup/gql'
 import * as faker from 'faker'
 import { getUsers, User } from '../support/mock/external/users'
