+++ frontend/cypress/page_objects/action.ts
@@ -1,6 +1,6 @@
 import { DropdownSelect, SideSheet } from '../page_objects/common'
-import { Action, Note } from '../support/mocks'
-import { FUSION_DATE_LOCALE } from '../support/helpers'
+import { Action, Note } from '../support/testsetup/mocks'
+import { FUSION_DATE_LOCALE } from '../support/helpers/helpers'
 import { Priority, Question } from '../../src/api/models'
 import { barrierToString, organizationToString } from '../../src/utils/EnumToString'
